+++ examples/scheduling-plain-text-email-example/src/main/java/com/test/PlainTextEmailSchedulingApplication.java
@@ -1,6 +1,5 @@
 package com.test;
 
-import it.ozimov.springboot.templating.mail.configuration.EnableEmailTools;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -13,19 +12,19 @@ import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 
 @SpringBootApplication
-@EnableEmailTools
-public class PlainTextEmailApplication {
+@ComponentScan(basePackages = {"com.test", "it.ozimov.springboot.templating.mail"})
+public class PlainTextEmailSchedulingApplication {
 
     @Autowired
     private TestService testService;
 
     public static void main(String[] args) {
-        SpringApplication.run(PlainTextEmailApplication.class, args);
+        SpringApplication.run(PlainTextEmailSchedulingApplication.class, args);
     }
 
     @PostConstruct
     public void sendEmail() throws UnsupportedEncodingException, InterruptedException {
-        testService.sendPlainTextEmail();
+        testService.scheduleSixEmails();
 
         close();
     }
@@ -38,7 +37,7 @@ public class PlainTextEmailApplication {
             }
         };
         Timer shutdownTimer = new Timer();
-        shutdownTimer.schedule(shutdownTask, TimeUnit.SECONDS.toMillis(3));
+        shutdownTimer.schedule(shutdownTask, TimeUnit.SECONDS.toMillis(20));
     }
 
 }