+++ examples/sending-mime-email-with-pebble-example/src/main/java/com/test/MimeEmailWithPebbleApplication.java
@@ -1,5 +1,6 @@
 package com.test;
 
+import it.ozimov.springboot.templating.mail.configuration.EnableEmailTools;
 import it.ozimov.springboot.templating.mail.service.exception.CannotSendEmailException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
@@ -13,7 +14,7 @@ import java.util.TimerTask;
 import java.util.concurrent.TimeUnit;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"com.test", "it.ozimov.springboot.templating.mail"})
+@EnableEmailTools
 public class MimeEmailWithPebbleApplication {
 
     @Autowired